秦始皇调整dialog为sheet，调整搜索歌词的入口

+++ app/src/main/java/com/ljyh/mei/data/repository/PlayerRepository.kt
@@ -43,8 +43,8 @@ class PlayerRepository(
         title: String,
         album: String,
         artist: String,
-        duration: Int,
-        id: Int
+        duration: Long,
+        id: Long
     ): Resource<LyricResult> {
         return withContext(Dispatchers.IO) {
             safeApiCall {
